refactor: InappBrowser 기능을 공통 파일로 분리
