+++ src/screens/contentDetail/components/linkDetail/LinkDetail.tsx
@@ -1,13 +1,11 @@
 import React, { useEffect, useState } from 'react'
 
-import { Platform, Linking, ImageURISource } from 'react-native'
-import { InAppBrowser, InAppBrowserOptions } from 'react-native-inappbrowser-reborn'
-import { InAppBrowseriOSOptions } from 'react-native-inappbrowser-reborn'
-import { InAppBrowserAndroidOptions } from 'react-native-inappbrowser-reborn'
+import { ImageURISource } from 'react-native'
 
 import { defaultImages } from '@/assets'
 import i18n from '@/locales'
 import { GetContentsResponse } from '@/models/Contents'
+import { openInappBrowser } from '@/services/InappBrowser'
 import { MetaData, getLinkOgTags } from '@/services/LinkService'
 
 import {
@@ -38,38 +36,12 @@ const LinkDetail = ({ content }: LinkDetailProps) => {
     }
   }, [content.link])
 
-  /**
-   * 인앱 브라우저를 엽니다.
-   */
-  const openInappBrowser = async () => {
-    try {
-      if (!content?.link) {
-        return
-      }
-
-      const isAvailable = await InAppBrowser.isAvailable()
-
-      if (isAvailable) {
-        InAppBrowser.open(content.link, getInAppBrowserOptions())
-        return
-      }
-
-      throw new Error('can not open inapp browser')
-    } catch (error) {
-      try {
-        Linking.openURL(content.link)
-      } catch (error) {
-        return
-      }
-    }
-  }
-
   if (!ogTags) {
     return <></>
   }
 
   return (
-    <Container onPress={() => openInappBrowser()}>
+    <Container onPress={() => openInappBrowser(content.link)}>
       <LinkPreview
         source={
           !isImageError && ogTags && ogTags['image']
@@ -90,24 +62,6 @@ const LinkDetail = ({ content }: LinkDetailProps) => {
       </TextContainer>
     </Container>
   )
-}
-
-/**
- * OS에 따른 InAppBrowserOptions을 가져옵니다.
- */
-const getInAppBrowserOptions = () => {
-  return Platform.select<InAppBrowserOptions>({
-    ios: {
-      dismissButtonStyle: 'close',
-      animated: true,
-      enableBarCollapsing: false,
-    } as InAppBrowseriOSOptions,
-    android: {
-      showTitle: false,
-      enableUrlBarHiding: true,
-      enableDefaultShare: true,
-    } as InAppBrowserAndroidOptions,
-  })
 }
 
 export default LinkDetail
